+++ docker/opendj/schedule_backup.sh
@@ -30,5 +30,6 @@ then
     echo "Scheduling incremental backup with cron schedule ${BACKUP_SCHEDULE_FULL}"
     bin/backup --backupDirectory ${BACKUP_DIRECTORY}  -p 4444 -D "cn=Directory Manager" \
     -j ${DIR_MANAGER_PW_FILE} --trustAll  --backupAll \
+    --incremental \
     --recurringTask "${BACKUP_SCHEDULE_INCREMENTAL}"
 fi